Give me change and Eight Queen Problem
A shopkeeper has coins of the following denominations: Rs. 5, Rs. 2, Re.
1,
50p, 25p, and 10p. When giving the change for his customers, he would
like to give the least number of coins. Write a program to help him
dispense the change.

In Chess, a queen can attack any piece that is horizontal, vertical or
diagonal to it. The challenge of 8 queens problem is to place 8 queens
on
a chessboard, so that no queen can attack another. One such solution is
shown above.
In this problem, given a board configuration, you will have to tell
whether a
queen will attack another. The positions are encoded as a string, each
character the column position, starting from the top row. For instance
the board is encoded as “52064713”.